Loginページのリファクタリング
