+++ src/components/organism/TitleFooter.tsx
@@ -1,0 +1,9 @@
+import { Box } from "@chakra-ui/react";
+
+export const TitleFooter = () => {
+	return (
+		<Box bg={"teal.100"} textAlign={"center"} h={"200px"}>
+			ふったー
+		</Box>
+	);
+};
