+++ src/router/Router.tsx
@@ -5,20 +5,11 @@ import { Login } from "../components/pages/Login";
 import Page404 from "../components/pages/Page404";
 import TopPage from "../components/pages/TopPage";
 import FooterLayout from "../components/templates/FooterLayout";
-import TitleHeaderLayout from "../components/templates/TitleHeaderLayout";
-//import HeaderLayout from "../components/templates/HeaderLayout";
 
 export const Router: VFC = (): ReactElement => {
 	return (
 		<Routes>
-			<Route
-				path="/login"
-				element={
-					<TitleHeaderLayout>
-						<Login />
-					</TitleHeaderLayout>
-				}
-			/>
+			<Route path="/login" element={<Login />} />
 			<Route
 				path="/top"
 				element={
